refactor: update project structure and configuration
- Renamed project from "with-vite" to "cloudcrafter" in bun.lock and package.json.
- Updated dependencies in package.json, including major version upgrades for fumadocs packages and added trusted dependencies.
- Replaced postcss.config.js with postcss.config.mjs for improved module support.
- Added new layout and configuration files for site structure.
- Removed deprecated components and styles, including layout files and global CSS.
- Introduced new assets for branding and updated homepage layout with a responsive design.
- Implemented a new source management system for documentation.
