--- frontend/apps/site/lib/source.ts
@@ -1,8 +1,0 @@
-import { docs, meta } from '@/.source'
-import { loader } from 'fumadocs-core/source'
-import { createMDXSource } from 'fumadocs-mdx'
-
-export const source = loader({
-	baseUrl: '/docs',
-	source: createMDXSource(docs, meta),
-})
