+++ frontend/apps/site/postcss.config.mjs
@@ -1,0 +1,5 @@
+export default {
+	plugins: {
+		'@tailwindcss/postcss': {},
+	},
+}
