+++ frontend/apps/site/src/app/docs/[[...slug]]/page.tsx
@@ -1,6 +1,5 @@
-import { MDXMermaid } from '@/components/docs/mdx/MDXMermaid'
 import { source } from '@/lib/source'
-import defaultMdxComponents from 'fumadocs-ui/mdx'
+import defaultMdxComponents, { createRelativeLink } from 'fumadocs-ui/mdx'
 import {
 	DocsBody,
 	DocsDescription,
@@ -9,11 +8,6 @@ import {
 } from 'fumadocs-ui/page'
 import { notFound } from 'next/navigation'
 
-const myComponents = {
-	...defaultMdxComponents,
-	Mermaid: MDXMermaid,
-}
-
 export default async function Page(props: {
 	params: Promise<{ slug?: string[] }>
 }) {
@@ -21,14 +15,21 @@ export default async function Page(props: {
 	const page = source.getPage(params.slug)
 	if (!page) notFound()
 
-	const MDX = page.data.body
+	const MDXContent = page.data.body
 
 	return (
 		<DocsPage toc={page.data.toc} full={page.data.full}>
 			<DocsTitle>{page.data.title}</DocsTitle>
 			<DocsDescription>{page.data.description}</DocsDescription>
 			<DocsBody>
-				<MDX components={{ ...myComponents }} />
+				<MDXContent
+					components={{
+						...defaultMdxComponents,
+						// this allows you to link to other pages with relative file paths
+						a: createRelativeLink(source, page),
+						// you can add other MDX components here
+					}}
+				/>
 			</DocsBody>
 		</DocsPage>
 	)
