+++ frontend/apps/site/src/app/global.css
@@ -1,0 +1,5 @@
+@import 'tailwindcss';
+@import 'fumadocs-ui/css/neutral.css';
+@import 'fumadocs-ui/css/preset.css';
+
+@source '../../../../node_modules/fumadocs-ui/dist/**/*.js';
