+++ frontend/apps/site/src/assets/logo.png
