+++ frontend/apps/site/src/lib/source.ts
@@ -1,0 +1,7 @@
+import { docs } from '@/.source'
+import { loader } from 'fumadocs-core/source'
+
+export const source = loader({
+	baseUrl: '/docs',
+	source: docs.toFumadocsSource(),
+})
