--- frontend/apps/site/styles/tailwind.css
@@ -1,12 +1,0 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-@keyframes move-x {
-  0% {
-    transform: translateX(var(--move-x-from));
-  }
-  100% {
-    transform: translateX(var(--move-x-to));
-  }
-}
