+++ frontend/apps/site/tsconfig.json
@@ -16,7 +16,8 @@
 		"jsx": "preserve",
 		"incremental": true,
 		"paths": {
-			"@/*": ["./*"]
+			"@/.source": ["./.source/index.ts"],
+			"@/*": ["./src/*"]
 		},
 		"plugins": [
 			{
