+++ frontend/package.json
@@ -11,7 +11,8 @@
 		"ui:add": "bun --cwd packages/ui ui:add",
 		"generate-clients": "bun --cwd packages/api generate-clients",
 		"generate-clients:e2e": "bun --cwd packages/e2e-testing generate-clients",
-		"generate-clients:all": "bun run generate-clients && bun run generate-clients:e2e",
+		"generate-clients:docs": "bun --cwd apps/site generate-clients",
+		"generate-clients:all": "bun run generate-clients && bun run generate-clients:e2e && bun run generate-clients:docs",
 		"format-and-lint": "biome check .",
 		"format-and-lint:fix": "biome check . --write",
 		"format-and-lint:hook": "biome check --write --staged --files-ignore-unknown=true --no-errors-on-unmatched  && git update-index --again",
@@ -28,5 +29,11 @@
 	},
 	"packageManager": "bun@1.0.35",
 	"name": "cloudcrafter",
-	"workspaces": ["apps/*", "packages/*"]
+	"workspaces": ["apps/*", "packages/*"],
+	"trustedDependencies": [
+		"@biomejs/biome",
+		"cloudcrafter-docs",
+		"core-js-pure",
+		"es5-ext"
+	]
 }
